Refactors to seperate client -> server actions
