+++ index.html
@@ -1,14 +1,13 @@
 <!DOCTYPE html>
 <html>
-    <head>
-        <meta charset="utf-8">
-        <title>TODO</title>
-        <script src="https://cdn.socket.io/socket.io-1.4.5.js"></script>
-    </head>
-    <body>
-        <input id="todo-input" type="text" placeholder="Feed the cat" autofocus>
-        <button type="button" onclick="add()">Make</button>
-        <ul id="todo-list"></ul>
-    </body>
-    <script src="client.js" async></script>
+  <head>
+    <meta charset="utf-8">
+    <title>TODO</title>
+  </head>
+  <body>
+    <input id="todo-input" type="text" placeholder="Feed the cat" autofocus>
+    <button type="button" onclick="add()">Make</button>
+    <ul id="todo-list"></ul>
+  </body>
+  <script src="client-bundle.js" async></script>
 </html>
