+++ package.json
@@ -18,8 +18,11 @@
   },
   "homepage": "https://github.com/adasupport/challenge#readme",
   "dependencies": {
+    "bufferutil": "^3.0.2",
     "express": "^4.14.0",
-    "socket.io": "^1.5.0",
-    "updeep": "^1.0.0"
+    "socket.io": "^2.0.3",
+    "socket.io-client": "^2.0.3",
+    "updeep": "^1.0.0",
+    "utf-8-validate": "^3.0.3"
   }
 }
